Strip dead CPP, clear -Widentities, and make the README suite run
Completes sized-grid-4c9.

base is pinned >=4.19, so the MIN_VERSION_base(4,11,0) guards around the
Data.Semigroup imports were unconditionally true. Drop the blocks and the
CPP pragma from Coord/Periodic.hs, Grid/Class.hs and tests/Test/Utils.hs.

Three fromIntegral calls were Integer -> Integer identities; removing them
makes the build clean at the -Wall -Wcompat -Widentities
-Wredundant-constraints level the cabal file already asks for.

Re-enable the readme test-suite, which had been commented out so long the
tutorial could rot unnoticed. It now runs instead of merely linking: the
documented entry point was an interactive clearScreen/getLine loop with
main = return (), so nothing was exercised. main now prints four
generations of the glider, covering applyRule, glider, gameOfLife and
displayGrid. That drops ansi-terminal along with four unused imports, and
removes five constraints GHC reports as redundant on the tutorial's own
glider signature.

Verified on GHC 9.12.3 and 9.14.1 via nix; both suites green.

+++ src/SizedGrid/Coord/HardWrap.hs
@@ -61,4 +61,4 @@ instance (1 <= n, KnownNat n) => AffineSpace (HardWrap n) where
   HardWrap a .+^ b = HardWrap $ fromJust $ numToOrdinal $
     max 0 $
     min (maxCoordSize (Proxy @(HardWrap n))) $
-    ((ordinalToNum a) + fromIntegral b)
+    (ordinalToNum a + b)
